impl RpcErrorExt
