+++ crates/transport/Cargo.toml
@@ -31,4 +31,4 @@ tracing.workspace = true
 wasm-bindgen-futures = "0.4"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-tokio = { workspace = true, features = ["rt"] }
+tokio = { workspace = true, features = ["rt", "time"] }
